Move GlobalContextRegistry nested modules
