+++ lib/rails_multitenant/global_context_registry.rb
@@ -1,5 +1,9 @@
 # frozen_string_literal: true
 
+require_relative 'global_context_registry/current'
+require_relative 'global_context_registry/current_instance'
+require_relative 'global_context_registry/registry_dependent_on'
+
 # Handles storing of global state that may be swapped out or unset at
 # various points. We use this in dandelion to store the current user,
 # org, catalog, etc.
@@ -8,206 +12,7 @@
 #
 module RailsMultitenant
   module GlobalContextRegistry
-
     extend self
-
-    module RegistryDependentOn
-
-      # Is this class dependent on changes in another GlobalContextRegistry-
-      # stored object? Register that dependency here.
-      def global_context_dependent_on(*klasses)
-        klasses.each { |klass| GlobalContextRegistry.send(:add_dependency, klass, self) }
-      end
-    end
-
-    # This module allows your to have a current, thread-local instance
-    # of this class. It currently assumes your class has a zero-arg
-    # constructor.
-    module Current
-      extend ActiveSupport::Concern
-
-      included do
-        class_attribute :default_provider, instance_writer: false
-      end
-
-      module ClassMethods
-        def current
-          GlobalContextRegistry.fetch(current_registry_obj) { __current_default }
-        end
-
-        def current=(object)
-          raise "#{object} is not a #{self}" if object.present? && !object.is_a?(self)
-
-          GlobalContextRegistry.set(current_registry_obj, object)
-          __clear_dependents!
-        end
-
-        def current?
-          GlobalContextRegistry.get(current_registry_obj).present?
-        end
-
-        def current!
-          current || raise("No current #{name} set")
-        end
-
-        def clear_current!
-          GlobalContextRegistry.delete(current_registry_obj)
-        end
-
-        def as_current(object)
-          old_object = current if current?
-          self.current = object
-          yield
-        ensure
-          self.current = old_object
-        end
-
-        include RegistryDependentOn
-
-        def provide_default(provider = nil, &block)
-          self.default_provider = provider ? provider.to_proc : block
-        end
-
-        private
-
-        def current_registry_obj
-          return @current_registry_obj if @current_registry_obj
-
-          @current_registry_obj = "#{__key_class.name.underscore}_obj".to_sym
-        end
-
-        def current_registry_default_provider
-          "#{__key_class.name.underscore}_default_provider".to_sym
-        end
-
-        def __current_default
-          if default_provider
-            default = default_provider.call(self)
-            raise "#{default} is not a #{self}" if default.present? && !default.is_a?(self)
-
-            default
-          end
-        end
-
-        def __clear_dependents!
-          GlobalContextRegistry.send(:dependencies_for, __key_class).each(&:clear_current!)
-        end
-
-        def __key_class
-          respond_to?(:base_class) ? base_class : self
-        end
-      end
-
-      def as_current
-        old_object = self.class.current if self.class.current?
-        self.class.current = self
-        yield
-      ensure
-        self.class.current = old_object
-      end
-
-      def current?
-        self.class.current? && equal?(self.class.current)
-      end
-
-    end
-
-    # This module allows you to have a current, thread-local instance
-    # of a class. This module assumes that you are mixing into a Rails
-    # model, and separately stores and id in thread local storage for
-    # lazy loading.
-    module CurrentInstance
-      extend ActiveSupport::Concern
-
-      module ClassMethods
-        def current_id=(id)
-          GlobalContextRegistry.delete(current_instance_registry_obj)
-          GlobalContextRegistry.set(current_instance_registry_id, id)
-          __clear_dependents!
-        end
-
-        def current=(object)
-          raise "#{object} is not a #{self}" if object.present? && !object.is_a?(self)
-
-          GlobalContextRegistry.set(current_instance_registry_obj, object)
-          GlobalContextRegistry.set(current_instance_registry_id, object.try(:id))
-          __clear_dependents!
-        end
-
-        def current_id
-          GlobalContextRegistry.get(current_instance_registry_id)
-        end
-
-        def current
-          GlobalContextRegistry.fetch(current_instance_registry_obj) do
-            (current_id ? find(current_id) : nil)
-          end
-        end
-
-        def current?
-          !!current
-        end
-
-        def current!
-          current || raise("No current #{name} set")
-        end
-
-        def as_current_id(id)
-          old_id = current_id
-          self.current_id = id
-          yield
-        ensure
-          self.current_id = old_id
-        end
-
-        def as_current(model)
-          old_model = current
-          self.current = model
-          yield
-        ensure
-          self.current = old_model
-        end
-
-        def clear_current!
-          GlobalContextRegistry.delete(current_instance_registry_obj)
-        end
-
-        private
-
-        def __clear_dependents!
-          key_class = respond_to?(:base_class) ? base_class : self
-          GlobalContextRegistry.send(:dependencies_for, key_class).each(&:clear_current!)
-        end
-
-        def current_instance_registry_id
-          return @current_instance_registry_id if @current_instance_registry_id
-
-          key_class = respond_to?(:base_class) ? base_class : self
-          @current_instance_registry_id = "#{key_class.name.underscore}_id".to_sym
-        end
-
-        def current_instance_registry_obj
-          return @current_instance_registry_obj if @current_instance_registry_obj
-
-          key_class = respond_to?(:base_class) ? base_class : self
-          @current_instance_registry_obj = "#{key_class.name.underscore}_obj".to_sym
-        end
-        include RegistryDependentOn
-      end
-
-      def as_current
-        old_id = self.class.current_id
-        self.class.current = self
-        yield
-      ensure
-        self.class.current_id = old_id
-      end
-
-      def current?
-        id == self.class.current_id
-      end
-
-    end
 
     # Set this global
     def set(symbol, value)
