+++ lib/rails_multitenant/global_context_registry/registry_dependent_on.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module RailsMultitenant
+  module GlobalContextRegistry
+    module RegistryDependentOn
+      # Is this class dependent on changes in another GlobalContextRegistry-
+      # stored object? Register that dependency here.
+      def global_context_dependent_on(*klasses)
+        klasses.each { |klass| GlobalContextRegistry.send(:add_dependency, klass, self) }
+      end
+    end
+  end
+end
